Added Application Theming
Added a themer class to allow us to apply a color scheme across the
entirety of the application.

Added custom themeable MEssage Boxes, and a default Dark Theme.

Added a message indicating the password lenght requirement to the
Account Creation form, and fixed a potential bug, which may have allowed
"NULL" usernames to be seen as unique.