Adds Icons to Mdi Forms, adds logo to Welcome Form.
Adds logos to all the subforms, and a big logo to the welcome page.

Fixes a typo on the welcome page, adds date labels to the Cash Flow
Form; the input box now returns a DialogBox Result, and fixes an error
when a user cancels updating/adding a profile.

Removed "add profile" from Account form.